:recycle: Checkout の Controller を結果と申請ページで分割
